refactor audio emitter system
